Extract layer variable mappings to a separate file
